+++ src/main/java/goal/jalal/goaljalal/GjlApplication.java
@@ -1,4 +1,4 @@
-package com.goaljalal.gjl;
+package goal.jalal.goaljalal;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
