+++ src/test/java/goal/jalal/goaljalal/GjlApplicationTests.java
@@ -1,4 +1,4 @@
-package com.goaljalal.gjl;
+package goal.jalal.goaljalal;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
